+++ codegen/src/rhai_module.rs
@@ -239,6 +239,7 @@ pub fn generate_body(
         gen_fn_tokens.push(quote! {
             #(#cfg_attrs)*
             #[allow(non_camel_case_types)]
+            #[doc(hidden)]
             pub struct #fn_token_name();
         });
 
@@ -250,6 +251,7 @@ pub fn generate_body(
             #[allow(unused_imports)]
             use super::*;
 
+            #[doc(hidden)]
             pub fn rhai_module_generate() -> Module {
                 let mut m = Module::new();
                 rhai_generate_into_module(&mut m, false);
@@ -257,6 +259,7 @@ pub fn generate_body(
                 m
             }
             #[allow(unused_mut)]
+            #[doc(hidden)]
             pub fn rhai_generate_into_module(m: &mut Module, flatten: bool) {
                 #(#set_fn_statements)*
                 #(#set_const_statements)*